enhance check
